% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplot.R
\name{plot.powerplot}
\alias{plot.powerplot}
\title{Plot power of multivariate RR methods}
\usage{
\method{plot}{powerplot}(x, ...)
}
\arguments{
\item{x}{a \code{\link{powerplot}} object}

\item{...}{ignored}
}
\description{
Plot estimated power from Monte Carlo simulation as a function of the sample size, separately for different effect sizes and multivariate RR methods
}
